Adding support for Int32 and Fixed (fixed length string/binary data)
Summary: SingleStore database types INT, CHAR(n), BINARY(n) are now supoorted by the client

Test Plan: ./build_and_test.sh lib data_types

Reviewers: dPoltorak-ua, amakarovych-ua

Reviewed By: dPoltorak-ua

Subscribers: engineering-list

JIRA Issues: PLAT-5429

Differential Revision: https://grizzly.internal.memcompute.com/D50251

+++ utils.cpp
@@ -45,27 +45,27 @@ namespace super_chunk
                     // integer types
                     case MYSQL_TYPE_TINY:
                     {
-                        column_info[i].type = BigInt;
+                        column_info[i].type = Int32;
                         break;
                     }
                     case MYSQL_TYPE_SHORT:
                     {
-                        column_info[i].type = BigInt;
+                        column_info[i].type = Int32;
                         break;
                     }
                     case MYSQL_TYPE_LONG:
                     {
-                        column_info[i].type = BigInt;
+                        column_info[i].type = Int32;
                         break;
                     }
                     case MYSQL_TYPE_INT24:
                     {
-                        column_info[i].type = BigInt;
+                        column_info[i].type = Int32;
                         break;
                     }
                     case MYSQL_TYPE_LONGLONG:
                     {
-                        column_info[i].type = BigInt;
+                        column_info[i].type = Int64;
                         break;
                     }
                     // floating point types
